requirements.txt
numpy
opencv_python
Pillow
timm
torch
torchsummary
torchvision
tensorboard
tqdm
